Clamp to 0/0/0
When a 256px tile at 0/0/0 is requested and the metatile size is 2, that leads to trying to request a metatile at -1/0/0, which doesn't exist. Instead, we clamp to 0/0/0, effectively returning the 512px 0/0/0 tile for 256px requests. This tile is at nominal zoom 1, which means that zoom 0 as a separate, styleable zoom.